+++ src/cache/inmemory/entityCache.ts
@@ -1,5 +1,6 @@
 import { NormalizedCache, NormalizedCacheObject, StoreObject } from './types';
 import { wrap, OptimisticWrapperFunction } from 'optimism';
+import { isReference } from './helpers';
 
 const hasOwn = Object.prototype.hasOwnProperty;
 
@@ -31,18 +32,22 @@ export abstract class EntityCache implements NormalizedCache {
   }
 
   public abstract addLayer(
-    id: string,
+    layerId: string,
     replay: (layer: EntityCache) => any,
   ): EntityCache;
 
-  public abstract removeLayer(id: string): EntityCache;
+  public abstract removeLayer(layerId: string): EntityCache;
 
   // Although the EntityCache class is abstract, it contains concrete
   // implementations of the various NormalizedCache interface methods that
   // are inherited by the Root and Layer subclasses.
 
   public toObject(): NormalizedCacheObject {
-    return this.data;
+    return { ...this.data };
+  }
+
+  public has(dataId: string): boolean {
+    return hasOwn.call(this.data, dataId);
   }
 
   public get(dataId: string): StoreObject {
@@ -53,12 +58,14 @@ export abstract class EntityCache implements NormalizedCache {
   public set(dataId: string, value: StoreObject): void {
     if (!hasOwn.call(this.data, dataId) || value !== this.data[dataId]) {
       this.data[dataId] = value;
+      delete this.refs[dataId];
       if (this.depend) this.depend.dirty(dataId);
     }
   }
 
   public delete(dataId: string): void {
-    this.data[dataId] = void 0;
+    delete this.data[dataId];
+    delete this.refs[dataId];
     if (this.depend) this.depend.dirty(dataId);
   }
 
@@ -77,6 +84,86 @@ export abstract class EntityCache implements NormalizedCache {
         this.set(dataId, newData[dataId]);
       });
     }
+  }
+
+  // Maps root entity IDs to the number of times they have been retained, minus
+  // the number of times they have been released. Retained entities keep other
+  // entities they reference (even indirectly) from being garbage collected.
+  private rootIds: {
+    [rootId: string]: number;
+  } = Object.create(null);
+
+  public retain(rootId: string): number {
+    return this.rootIds[rootId] = (this.rootIds[rootId] || 0) + 1;
+  }
+
+  public release(rootId: string): number {
+    if (this.rootIds[rootId] > 0) {
+      const count = --this.rootIds[rootId];
+      if (!count) delete this.rootIds[rootId];
+      return count;
+    }
+    return 0;
+  }
+
+  // This method will be overridden in the Layer class to merge root IDs for all
+  // layers (including the root).
+  public getRootIdSet() {
+    return new Set(Object.keys(this.rootIds));
+  }
+
+  // The goal of garbage collection is to remove IDs from the Root layer of the
+  // cache that are no longer reachable starting from any IDs that have been
+  // explicitly retained (see retain and release, above). Returns an array of
+  // dataId strings that were removed from the cache.
+  public gc() {
+    const ids = this.getRootIdSet();
+    const snapshot = this.toObject();
+    ids.forEach(id => {
+      if (hasOwn.call(snapshot, id)) {
+        // Because we are iterating over an ECMAScript Set, the IDs we add here
+        // will be visited in later iterations of the forEach loop only if they
+        // were not previously contained by the Set.
+        Object.keys(this.findChildRefIds(id)).forEach(ids.add, ids);
+        // By removing IDs from the snapshot object here, we protect them from
+        // getting removed from the root cache layer below.
+        delete snapshot[id];
+      }
+    });
+    const idsToRemove = Object.keys(snapshot);
+    if (idsToRemove.length) {
+      let root: EntityCache = this;
+      while (root instanceof Layer) root = root.parent;
+      idsToRemove.forEach(root.delete, root);
+    }
+    return idsToRemove;
+  }
+
+  // Lazily tracks { __ref: <dataId> } strings contained by this.data[dataId].
+  private refs: {
+    [dataId: string]: Record<string, true>;
+  } = Object.create(null);
+
+  public findChildRefIds(dataId: string): Record<string, true> {
+    if (!hasOwn.call(this.refs, dataId)) {
+      const found = this.refs[dataId] = Object.create(null);
+      const workSet = new Set([this.data[dataId]]);
+      // Within the cache, only arrays and objects can contain child entity
+      // references, so we can prune the traversal using this predicate:
+      const canTraverse = (obj: any) => obj !== null && typeof obj === 'object';
+      workSet.forEach(obj => {
+        if (isReference(obj)) {
+          found[obj.__ref] = true;
+        } else if (canTraverse(obj)) {
+          Object.values(obj)
+            // No need to add primitive values to the workSet, since they cannot
+            // contain reference objects.
+            .filter(canTraverse)
+            .forEach(workSet.add, workSet);
+        }
+      });
+    }
+    return this.refs[dataId];
   }
 }
 
@@ -107,14 +194,14 @@ export namespace EntityCache {
     }
 
     public addLayer(
-      id: string,
+      layerId: string,
       replay: (layer: EntityCache) => any,
     ): EntityCache {
       // The replay function will be called in the Layer constructor.
-      return new Layer(id, this, replay, this.sharedLayerDepend);
+      return new Layer(layerId, this, replay, this.sharedLayerDepend);
     }
 
-    public removeLayer(): Root {
+    public removeLayer(layerId: string): Root {
       // Never remove the root layer.
       return this;
     }
@@ -125,9 +212,9 @@ export namespace EntityCache {
 // of the EntityCache.Root class.
 class Layer extends EntityCache {
   constructor(
-    private id: string,
-    private parent: EntityCache,
-    private replay: (layer: EntityCache) => any,
+    public readonly id: string,
+    public readonly parent: Layer | EntityCache.Root,
+    public readonly replay: (layer: EntityCache) => any,
     public readonly depend: DependType,
   ) {
     super();
@@ -135,17 +222,17 @@ class Layer extends EntityCache {
   }
 
   public addLayer(
-    id: string,
+    layerId: string,
     replay: (layer: EntityCache) => any,
   ): EntityCache {
-    return new Layer(id, this, replay, this.depend);
+    return new Layer(layerId, this, replay, this.depend);
   }
 
-  public removeLayer(id: string): EntityCache {
+  public removeLayer(layerId: string): EntityCache {
     // Remove all instances of the given id, not just the first one.
-    const parent = this.parent.removeLayer(id);
+    const parent = this.parent.removeLayer(layerId);
 
-    if (id === this.id) {
+    if (layerId === this.id) {
       // Dirty every ID we're removing.
       // TODO Some of these IDs could escape dirtying if value unchanged.
       if (this.depend) {
@@ -168,6 +255,26 @@ class Layer extends EntityCache {
     };
   }
 
+  public has(dataId: string): boolean {
+    // Because the Layer implementation of the delete method uses void 0 to
+    // indicate absence, that's what we need to check for here, rather than
+    // calling super.has(dataId).
+    if (hasOwn.call(this.data, dataId) && this.data[dataId] === void 0) {
+      return false;
+    }
+    return this.parent.has(dataId);
+  }
+
+  public delete(dataId: string): void {
+    super.delete(dataId);
+    // In case this.parent (or one of its ancestors) has an entry for this ID,
+    // we need to shadow it with an undefined value, or it might be inherited
+    // by the Layer#get method.
+    this.data[dataId] = void 0;
+  }
+
+  // All the other inherited accessor methods work as-is, but the get method
+  // needs to fall back to this.parent.get when accessing a missing dataId.
   public get(dataId: string): StoreObject {
     if (hasOwn.call(this.data, dataId)) {
       return super.get(dataId);
@@ -182,6 +289,22 @@ class Layer extends EntityCache {
       this.depend(dataId);
     }
     return this.parent.get(dataId);
+  }
+
+  // Return a Set<string> of all the ID strings that have been retained by this
+  // Layer *and* any layers/roots beneath it.
+  public getRootIdSet(): Set<string> {
+    const ids = this.parent.getRootIdSet();
+    super.getRootIdSet().forEach(ids.add, ids);
+    return ids;
+  }
+
+  public findChildRefIds(dataId: string): Record<string, true> {
+    const fromParent = this.parent.findChildRefIds(dataId);
+    return hasOwn.call(this.data, dataId) ? {
+      ...fromParent,
+      ...super.findChildRefIds(dataId),
+    } : fromParent;
   }
 }