+++ src/cache/inmemory/writeToStore.ts
@@ -76,6 +76,12 @@ export class StoreWriter {
     dataIdFromObject?: IdGetter;
   }): NormalizedCache {
     const operationDefinition = getOperationDefinition(query)!;
+
+    // Any IDs written explicitly to the cache (including ROOT_QUERY, most
+    // frequently) will be retained as reachable root IDs on behalf of their
+    // owner DocumentNode objects, until/unless evicted for all owners.
+    store.retain(dataId);
+
     return this.writeSelectionSetToStore({
       result,
       dataId,
@@ -93,7 +99,7 @@ export class StoreWriter {
     });
   }
 
-  public writeSelectionSetToStore({
+  private writeSelectionSetToStore({
     result,
     dataId,
     selectionSet,